Full docker app (#80)
* Added a docker container for the application

* Finalize Dockerized app

+++ Makefile
@@ -1,22 +1,60 @@
-TOOLS = docker-compose run --rm tools
-CONSOLE = ./bin/console
+DOCKER:=$(shell docker-compose -v dot 2> /dev/null)
 
-all: boot install assets run
+ifdef DOCKER
+    APP=docker-compose exec -T app
+    TOOLS=docker-compose run --rm tools
+endif
 
-boot:
+CONSOLE=$(APP) bin/console
+
+.PHONY: clean
+
+all: start dependencies db perm
+
+dependencies: composer yarn assets perm
+
+stop:
+ifdef DOCKER
+	docker-compose down
+endif
+
+start:
+ifdef DOCKER
 	docker-compose up -d
+else
+	@bash -c "if [ -a vendor/autoload.php ]; then $(CONSOLE) server:run; fi;"
+endif
 
-install:
-	$(TOOLS) sh -c "composer install && yarn install"
+composer:
+	$(APP) composer install
+
+yarn:
+	$(TOOLS) yarn install
+
+db:
+	$(APP) php -r "for(;;){if(@fsockopen('db',3306)){break;}}" # Wait for MariaDB
+	$(CONSOLE) doctrine:database:drop --force --if-exists
+	$(CONSOLE) doctrine:database:create --if-not-exists
+	$(CONSOLE) doctrine:schema:create
+	$(CONSOLE) doctrine:fixtures:load -n
+	$(CONSOLE) app:content:prepare
+
+clean:
+	$(CONSOLE) cache:clear --no-warmup
+	$(CONSOLE) cache:clear --no-warmup --env=prod
+	$(CONSOLE) cache:clear --no-warmup --env=test
+	$(APP) rm -rf var/logs/*
+	$(APP) rm -rf var/sessions/*
+	$(APP) rm -rf web/built/*
+
+perm:
+	$(APP) chmod 777 -R var
 
 assets:
-	$(TOOLS) sh -c "npm run build-dev"
-
-run:
-	$(CONSOLE) server:run
+	$(TOOLS) npm run build-dev
 
 watch:
 	$(TOOLS) npm run watch
 
 test:
-	$(TOOLS) ./vendor/bin/simple-phpunit
+	$(APP) vendor/bin/simple-phpunit
